requirements.txt
tensorflow==2.14.0
numpy==1.23.5
matplotlib==3.7.1
opencv-python==4.8.0
Pillow==10.0.0 